Make import-gnulib.sh faster when neither import-gnulib.config nor gnulib has changed.
* import-gnulib.sh (gnulib_changed): New variable.
(do_checkout): set gnulib_changed=true if the gnulib code didn't
already correspond to our configured version.  Don't use a large
subshell, because we need to set $gnulib_changed and don't want
the value to be discarded.
(main): Run gnulib-tool if gnulib was not at the correct version
or the import-gnulib.config file has changed.

Signed-off-by: James Youngman <jay@gnu.org>