find: use FTS_VERBATIM (and update gnulib)
The recent change v0.0-7611-g3a9002d in gnulib that made fts_open strip
trailing slashes from input path names had a negative impact on
findutils that relies on the old fts_open behavior to implement POSIX
requirement that each path operand of the find utility shall be
evaluated unaltered as it was provided, including all trailing slash
characters.

A new bit flag FTS_VERBATIM was added later as the option to disable
that stripping when it is not desirable.

* gnulib: update to latest, to get FTS_VERBATIM support.
* find/ftsfind.c (ftsoptions): Set the FTS_VERBATIM bit.

As a result of the gnulib update, re-generate doc/regexprops.texi.

* doc/regexprops.texi: re-generate to describe the changes to
gnulib for the regular expression types "awk" (to add character
class support), "gnu-awk" (to add range support) and "posix-awk"
(also to add range suport).
* NEWS: Describe this change.