Several updates and clarifications to README-hacking.
* README-hacking: Explain that tool-versions.txt is not checked
into git.  The repository has grown from 14MB to 16MB.  Gnulib is
now a submodule and the directory we keep it in has changed.
Remove the suggestion to rsync the translations from the
translation project, since the files that fetches seem to be out
of date.  Move the instruction to update ChangeLog to a more
prominent place.  Mention that there is an electronic process for
assignments in some locations.  Update the copyright year.