maint: generate ChangeLog from git commit messages
* ChangeLog: Remove, as it is not maintained manually any longer.
* Makefile.am (gen-ChangeLog): Add phony target to create the ChangeLog
file in the distribution tarball from the git commit messages - starting
from 2014 - and concatenating the file 'ChangeLog-2013'.
(dist-hook): Add the above new target as dependency.
* ChangeLog-2013: Add a newline at the top for nicer separation in
the concatenated ChangeLog file in the distribution tarball.
Bump Copyright year.
* import-gnulib.config (maintainer_modules): Add 'gitlog-to-changelog'
module.  Furthermore create an empty ChangeLog file to silence automake,
and set DO_NOT_WANT_CHANGELOG_DRIVER environment variable as we do not
want the import to fail if the changelog driver is not installed.
* build-aux/git-log-fix: Add rules to fix some git commit messages
during ChangeLog generation.
* build-aux/.gitignore: Add entry for 'gitlog-to-changelog'.
* build-aux/.cvsignore: Likewise.
* .gitignore: Likewise.
* README-hacking: Remove item to add a ChangeLog entry, and instead
mention to give a decent commit message.  In the command sequence for
a complete session, remove ChangeLog editing, too.
* doc/find-maint.texi (Release Information): Remove ChangeLog editing.
*

Signed-off-by: James Youngman <jay@gnu.org>