Don't include dblocation.texi from original spot, symlink it.
* doc/Makefile.am (dblocation.texi): Create this as a symbolic
link.  We do this because Automake now invokes texi2dvi in "tidy"
mode and hence changes the working directory.  That means that
relative path includes no longer works in Texinfo files, so we
need to avoid doing that.  Avoid distributing dblocation.texi
(since tar would dereference the symbolic link when making the
source archive, since the -h option is passed).
* doc/find.texi: Include dblocation.texi instead of
../locate/dblocation.texi.
* configure.ac: use AC_PROG_LN_S.
* locate/Makefile.am (EXTRA_DIST): Don't distribute dblocation.texi.