Print the modification time of the database for --statistics.
* locate/locate.c: Include stat-time.h.
(print_stats): Add an argument database_mtime, indicating the
modification time of the database.  Print that with the rest of
the statistics.
(search_one_database): Accept an additional argument,
database_mtime, and pass it to print_stats.
(dolocate): Extract the mtime of the database file and pass it to
search_one_database.