Documentation fix for bug #36960 and greater effect of POSIXLY_CORRECT.
* find/parser.c (parse_table): -warn is now parsed as a positional
option instead of a global one. This does not affect the effect of
-warn.  (found_parser): Warnings are no longer issued for global
options specified in confusing positions when the POSIXLY_CORRECT
environment variable is set.  POSIXLY_CORRECT also suppresses warnings
about -d and the use of '/' in the argument to -name.  * find/find.1:
Eliminate the duplicate section name OPTIONS, using the OPTIONS
section just for -L, -P and so on.  Other options are now documented
in the subsection POSITIONAL OPTIONS (if they take effect only on
items on the command line which appear later) or GLOBAL OPTIONS (if
they take effect even on tests/actions appearing before them).  The
-regextype option in particular is now correctly documented as being
positional.  We rename the EXPRESSIONS section to EXPRESSION in order
to match the usage summary.  The OPERATORS heading is demoted to a
subheading so that it fits within EXPRESSION.  (POSITIONAL OPTIONS):
be more explicit about how warnings are controlled by -warn, -nowarn
and POSIXLY_CORRECT.  * NEWS: Describe the changes to the find manual
page and warnings.